+++ packages/apptive_grid_core/lib/model/grid/grid_row.dart
@@ -15,7 +15,6 @@ class GridRow {
   factory GridRow.fromJson(
     dynamic json,
     List<GridField> fields,
-    dynamic schema,
   ) {
     final data = json['fields'] as List;
     final entries = List<GridEntry>.generate(
@@ -23,7 +22,7 @@ class GridRow {
       (i) => GridEntry.fromJson(
         data[i],
         fields[i],
-        schema['properties']['fields']['items'][i],
+        fields[i].schema,
       ),
     );
     return GridRow(